+++ src/ListContainer.test.jsx
@@ -2,22 +2,19 @@ import { render } from '@testing-library/react';
 
 import { useSelector } from 'react-redux';
 
-import App from './App';
+import ListContainer from './ListContainer';
 
 jest.mock('react-redux');
 
-test('App', () => {
+test('ListContainer', () => {
   useSelector.mockImplementation((selector) => selector({
-    taskTitle: '',
     tasks: [
       { id: 1, title: '아무 것도 하지 않기 #1' },
       { id: 2, title: '아무 것도 하지 않기 #2' },
     ],
   }));
 
-  const { getByText } = render(<App />);
-
-  expect(getByText(/추가/)).not.toBeNull();
+  const { getByText } = render(<ListContainer />);
 
   expect(getByText(/아무 것도 하지 않기 #1/)).not.toBeNull();
 });
